+++ src/components/UI/Layout/ThreePanelLayout/MainPanel/MainPanel.tsx
@@ -4,23 +4,18 @@ import {ReactNode, useState} from "react";
 import {AnimatePresence, motion} from 'framer-motion';
 import StyledPanel from "../../../Common/Panel/Panel";
 import {screenWidth} from '../../../screenSizes';
+import useAppState from "../../../../../store/appState/useAppState";
 
 type MainPanelProps = {
-    children: ReactNode,
-    showLeftHandler: () => void
+    rightPanel: ReactNode,
+    children: ReactNode
 }
 
 const StyledMainPanel = styled(StyledPanel)`
-  background-color: green;
-  flex: 1;
-  max-width: unset;
   display: flex;
-  align-items: center;
-  justify-content: space-between;
-  position: absolute;
-  @media (${screenWidth.narrowWidth}) {
-    position: relative;
-  }
+  width: 100%;
+  flex: 1;
+  content: '';
 `
 
 const StyledMainContents = styled(motion.div)`
@@ -28,29 +23,13 @@ const StyledMainContents = styled(motion.div)`
   flex: 1;
 `
 
-const MainPanel = ({children, showLeftHandler}: MainPanelProps) => {
-    const [showRight, setShowRight] = useState(false)
-    const [showMain, setShowMain] = useState(false)
+const MainPanel = ({children, rightPanel}: MainPanelProps) => {
     return (
         <StyledMainPanel>
-            <button style={{position: 'absolute', top: 100, left: 0, zIndex: 100000}}
-                    onClick={() => showLeftHandler()}>
-                SHOW LEFT
-            </button>
-            <button style={{position: 'absolute', top: 0, left: 0, zIndex: 100000}}
-                    onClick={() => setShowRight(!showRight)}>
-                SHOW RIGHT
-            </button>
-            <button style={{position: 'absolute', top: 50, left: 0, zIndex: 100000}}
-                    onClick={() => setShowMain(!showMain)}>
-                SHOW MAIN
-            </button>
-            <StyledMainContents layout key='main-contents'>
-                Main panel
+            <StyledMainContents>
+                {children}
             </StyledMainContents>
-            {
-                showRight && children
-            }
+            {rightPanel}
         </StyledMainPanel>
     )
 }