+++ src/contexts/pageNavigator/pageNavigatorActions.ts
@@ -1,0 +1,10 @@
+import {NavigatePageAction} from "./pageNavigatorReducer"
+
+export const pageNavigator = {
+    navigateToPage: (pageName: string): NavigatePageAction => {
+        return {
+            type: 'navigateToPage',
+            payload: pageName
+        }
+    }
+}